// frontend/src/api/models/ServiceOrderItemResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Clinic Management API
 * API documentation for the Clinic Management system
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ServiceCatalogResponse } from './ServiceCatalogResponse';
import {
    ServiceCatalogResponseFromJSON,
    ServiceCatalogResponseFromJSONTyped,
    ServiceCatalogResponseToJSON,
    ServiceCatalogResponseToJSONTyped,
} from './ServiceCatalogResponse';

/**
 * 
 * @export
 * @interface ServiceOrderItemResponse
 */
export interface ServiceOrderItemResponse {
    /**
     * 
     * @type {string}
     * @memberof ServiceOrderItemResponse
     */
    id?: string;
    /**
     * 
     * @type {ServiceCatalogResponse}
     * @memberof ServiceOrderItemResponse
     */
    service?: ServiceCatalogResponse;
    /**
     * 
     * @type {number}
     * @memberof ServiceOrderItemResponse
     */
    price?: number;
    /**
     * 
     * @type {string}
     * @memberof ServiceOrderItemResponse
     */
    note?: string;
}

/**
 * Check if a given object implements the ServiceOrderItemResponse interface.
 */
export function instanceOfServiceOrderItemResponse(value: object): value is ServiceOrderItemResponse {
    return true;
}

export function ServiceOrderItemResponseFromJSON(json: any): ServiceOrderItemResponse {
    return ServiceOrderItemResponseFromJSONTyped(json, false);
}

export function ServiceOrderItemResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ServiceOrderItemResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'service': json['service'] == null ? undefined : ServiceCatalogResponseFromJSON(json['service']),
        'price': json['price'] == null ? undefined : json['price'],
        'note': json['note'] == null ? undefined : json['note'],
    };
}

export function ServiceOrderItemResponseToJSON(json: any): ServiceOrderItemResponse {
    return ServiceOrderItemResponseToJSONTyped(json, false);
}

export function ServiceOrderItemResponseToJSONTyped(value?: ServiceOrderItemResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'service': ServiceCatalogResponseToJSON(value['service']),
        'price': value['price'],
        'note': value['note'],
    };
}